requirements.txt
# Python
python==3.9.12
# Django y herramientas
django==4.1
djangorestframework==3.13.1
# Filtros
django_filter==22.1
